refactor: split sdk
